requirements.txt
flask
numpy
dicttoxml
requests
google-cloud-storage
gunicorn==19.6.0
opencv-python-headless
